refactor: move DebugImage and DebugMeta to Protocol
