+++ src/Sentry/Internal/Extensions/PEReaderExtensions.cs
@@ -1,4 +1,5 @@
 using Sentry.Extensibility;
+using Sentry.Protocol;
 
 namespace Sentry.Internal.Extensions;
