+++ src/Sentry/Internal/ILSpy/SingleFileApp.cs
@@ -1,5 +1,6 @@
 using Sentry.Extensibility;
 using Sentry.Internal.Extensions;
+using Sentry.Protocol;
 
 namespace Sentry.Internal.ILSpy;
