+++ src/Sentry/Platforms/Native/CFunctions.cs
@@ -1,5 +1,6 @@
 using Sentry.Extensibility;
 using Sentry.Internal.Extensions;
+using Sentry.Protocol;
 
 namespace Sentry.Native;
